conexão DB
ajuste na conexão com novo modelo de mvc
