+++ main.py
@@ -1,3 +1,3 @@
 if __name__ == "__main__":
-    from App import desktop
-    # from App import web
+    #from App import desktop
+    from App import web
